// Cosmos.Clinical/src/app/models/userrrolemanagement/studyidtitleview.model.ts
export class StudyIdTitleViewModel {
  sponsorSiteStudyCdaInvitationId: number;
  sponsorStudyInfoId: number;
  title: string;
  protocolNumber: string;
  siteInfoId: number;
  investigatorInfoId: number;
  siteNumber: string;
  piFirstName: string;
  piLastName: string;
  piMiddleName: string;
  salutationId: number;
  salutationTitle: string;
  sponsorName: string;
  isBlind: boolean;
  roles: number[];
}